Add option to specify number of feature to consider; fix several bugs.
Maximum depth has been re-interpreted as maximum distance from the
root. This effectively adds one extra layer of nodes, to allow 1:1
comparison with the mk3 trainer.

When determining whether or not to include a feature, the logic was
inverted. That is, a succesful coin flip should have implied that
the corresponding feature would be considered, but instead caused
the feature to be ignored.

In the majority voting implementation used during classification,
the label should be true only when the number of votes is strictly
larger than half of the total number of voters, not larger than or
equal.