Fix classification where the number of features is smaller than the number of classes.
The vote table was created with shape #points x #features, while this
should have been #points x #classes. Otherwise, for label values larger
than the number of features, an out of bounds write is performed.

The number of classes distinguished by the classifier was not stored
in the model file, so fixing this issue meant changing the model file
format. Consequently, code related to reading and writing of the model
file needed to be updated as well.