Modify balsa_test to not use mkdtemp() which is POSIX.
Instead of storing temporary files in a temporary directory with
a unique name, temporary files are now created in the working
directory with non-unique, hard-coded names. This would only lead
to issues when multiple instances of balsa_test are started
simultaneously in the same directory. While this is possible, it
is not a realistic scenario.