Remove feature count as parameter for classify(AndVote)().
The feature count of the classifier will be used. It will
be checked that the number of data points passed to these
functions is an integral multiple of the number of
features. This check is only approximate and can fail in
several ways. In general, the assumption is that the caller
will pass well-formed arrays to these functions. The check
is merely an extra measure to trap some glaring mistakes.